Create the NuGet release tool
