+++ NuGetReleaseTool/NuGetReleaseTool.sln
@@ -1,0 +1,22 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.31903.59
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NuGetReleaseTool", "NuGetReleaseTool\NuGetReleaseTool.csproj", "{61E2CA24-55B2-4FF2-B8DB-70F1FD12C082}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{61E2CA24-55B2-4FF2-B8DB-70F1FD12C082}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{61E2CA24-55B2-4FF2-B8DB-70F1FD12C082}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{61E2CA24-55B2-4FF2-B8DB-70F1FD12C082}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{61E2CA24-55B2-4FF2-B8DB-70F1FD12C082}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
