+++ NuGetReleaseTool/NuGetReleaseTool/GitCredentials.cs
