+++ NuGetReleaseTool/NuGetReleaseTool/NuGetReleaseTool.csproj
@@ -1,0 +1,14 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+	<PropertyGroup>
+		<OutputType>Exe</OutputType>
+		<TargetFramework>net7.0</TargetFramework>
+		<ImplicitUsings>enable</ImplicitUsings>
+		<Nullable>enable</Nullable>
+	</PropertyGroup>
+
+	<ItemGroup>
+		<PackageReference Include="CommandLineParser" Version="2.7.82" />
+		<PackageReference Include="octokit" Version="0.42.0" />
+	</ItemGroup>
+</Project>
