+++ NuGetReleaseTool/NuGetReleaseTool/ReleaseGenerator/IssueType.cs
