--- ReleaseNotesGenerator/ReleaseNotesGenerator.sln
@@ -1,30 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.4.33015.44
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ReleaseNotesGenerator", "ReleaseNotesGenerator\ReleaseNotesGenerator.csproj", "{95B80A75-AF6E-424A-880A-3EFA4FDF26F4}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{AAACF866-0617-416C-AE81-9644BB2868EF}"
-	ProjectSection(SolutionItems) = preProject
-		README.md = README.md
-	EndProjectSection
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{95B80A75-AF6E-424A-880A-3EFA4FDF26F4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{95B80A75-AF6E-424A-880A-3EFA4FDF26F4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{95B80A75-AF6E-424A-880A-3EFA4FDF26F4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{95B80A75-AF6E-424A-880A-3EFA4FDF26F4}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {4EC09464-4801-4B1B-99AB-FD661A283DF7}
-	EndGlobalSection
-EndGlobal
