--- ReleaseNotesGenerator/ReleaseNotesGenerator/Properties/AssemblyInfo.cs
@@ -1,11 +1,0 @@
-﻿using System.Reflection;
-using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
-
-// Setting ComVisible to false makes the types in this assembly not visible 
-// to COM components.  If you need to access a type in this assembly from 
-// COM, set the ComVisible attribute to true on that type.
-[assembly: ComVisible(false)]
-
-// The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("95b80a75-af6e-424a-880a-3efa4fdf26f4")]
