--- ReleaseNotesGenerator/ReleaseNotesGenerator/ReleaseNotesGenerator.csproj
@@ -1,10 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net6.0</TargetFramework>
-  </PropertyGroup>
-  <ItemGroup>
-    <PackageReference Include="CommandLineParser" Version="2.7.82" />
-    <PackageReference Include="octokit" Version="0.42.0" />
-  </ItemGroup>
-</Project>
